MR-35: ASYNC/AWAIT POST/DELETE methods

+++ src/controllers/member.js
@@ -53,34 +53,45 @@ const getMembersById = (req, res) => {
   }
 };
 
-const createMember = (req, res) => {
+const createMember = async (req, res) => {
   const {
     firstName, lastName, dni, phone, email, city, birthday, postalCode, isActive, membership,
   } = req.body;
 
-  Member.create({
-    firstName,
-    lastName,
-    dni,
-    phone,
-    email,
-    city,
-    birthday,
-    postalCode,
-    isActive,
-    membership,
-  })
-
-    .then((result) => res.status(201).json({
-      message: `Member ${result.id} was created successfully!`,
-      data: result,
+  try {
+    const found = await Member.findOne({ dni });
+    if (found) {
+      throw Error('This member already exist');
+    }
+    const createdMember = await Member.create({
+      firstName,
+      lastName,
+      dni,
+      phone,
+      email,
+      city,
+      birthday,
+      postalCode,
+      isActive,
+      membership,
+    });
+    return res.status(201).json({
+      message: `Member ${createdMember.id} was created successfully!`,
+      data: createdMember,
       error: false,
-    }))
-
-    .catch((error) => res.status(400).json({
-      message: 'Failed to create an Member',
+    });
+  } catch (error) {
+    if (error.message === 'This member already exist') {
+      return res.status(400).json({
+        message: 'Failed to create an Member, member already exist',
+        error,
+      });
+    }
+    return res.status(500).json({
+      message: 'Error was ocurred',
       error,
-    }));
+    });
+  }
 };
 
 const editMember = (req, res) => {
@@ -94,7 +105,6 @@ const editMember = (req, res) => {
       message: `ID: ${id} format is not valid`,
     });
   }
-
   Member.findByIdAndUpdate(
     id,
     {
@@ -130,31 +140,50 @@ const editMember = (req, res) => {
     }));
 };
 
-const deleteMember = (req, res) => {
+const deleteMember = async (req, res) => {
   const { id } = req.params;
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    res.status(404).json({
-      message: `ID: ${id} is not valid`,
+
+  try {
+    if (!id) {
+      throw Error('Blank ID');
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw Error('Invalid');
+    }
+    const memberDeleted = await Member.findByIdAndDelete(id);
+    if (!memberDeleted) {
+      throw Error('Member not found');
+    }
+    return res.status(200).json({
+      message: 'Member was deleted succesfully!',
+      data: memberDeleted,
+      error: false,
+    });
+  } catch (error) {
+    if (error.message === 'Blank ID') {
+      return res.status(400).json({
+        message: 'Please enter an ID',
+        error,
+      });
+    }
+    if (error.message === 'Invalid ID') {
+      return res.status(400).json({
+        message: `${id} is not a valid ID`,
+        error,
+      });
+    }
+    if (error.message === 'Member not found') {
+      return res.status(400).json({
+        message: 'Member not found',
+        error,
+      });
+    }
+    return res.status(500).json({
+      message: 'Error was ocurred',
+      error,
     });
   }
-  Member.findByIdAndDelete(id)
-    .then((result) => {
-      if (!result) {
-        return res.status(404).json({
-          message: `Member with ID: ${id} was not found`,
-        });
-      }
-      return res.status(200).json({
-        message: 'Member was deleted succesfully!',
-        data: result,
-        error: false,
-      });
-    })
-    .catch((error) => res.status(400).json({
-      message: 'Request error',
-      data: error,
-      error,
-    }));
 };
 
 module.exports = {
